refactor: repository/exception 패키지 내 도메인별 패키지 통합 관리
