+++ src/main/java/wooteco/subway/repository/exception/DuplicateLineColorException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.repository.exception.line;
+package wooteco.subway.repository.exception;
 
 public class DuplicateLineColorException extends IllegalArgumentException {
