+++ src/main/java/wooteco/subway/repository/exception/NoSuchLineException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.repository.exception.line;
+package wooteco.subway.repository.exception;
 
 import java.util.NoSuchElementException;
