+++ src/test/java/wooteco/subway/repository/dao/SectionDaoTest.java
@@ -9,7 +9,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 
 import javax.sql.DataSource;
-import wooteco.subway.repository.dao.SectionDao;
 import wooteco.subway.repository.dao.entity.section.SectionEntity;
 
 @JdbcTest
