+++ src/test/java/wooteco/subway/repository/dao/StationDaoTest.java
@@ -14,7 +14,6 @@ import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 
 import javax.sql.DataSource;
 import wooteco.subway.domain.station.Station;
-import wooteco.subway.repository.dao.StationDao;
 import wooteco.subway.repository.dao.entity.EntityAssembler;
 import wooteco.subway.repository.dao.entity.station.StationEntity;
