+++ src/test/java/wooteco/subway/service/LineServiceTest.java
@@ -20,8 +20,8 @@ import wooteco.subway.repository.SubwayRepository;
 import wooteco.subway.repository.dao.LineDao;
 import wooteco.subway.repository.dao.SectionDao;
 import wooteco.subway.repository.dao.StationDao;
-import wooteco.subway.repository.exception.line.DuplicateLineColorException;
-import wooteco.subway.repository.exception.line.DuplicateLineNameException;
+import wooteco.subway.repository.exception.DuplicateLineColorException;
+import wooteco.subway.repository.exception.DuplicateLineNameException;
 import wooteco.subway.service.dto.line.LineRequest;
 import wooteco.subway.service.dto.line.LineResponse;
 import wooteco.subway.service.dto.line.LineUpdateRequest;
